Recactor authentication and add aphrodite
