+++ examples/forum-relay/public/index.html
@@ -4,6 +4,7 @@
     <meta charset="utf-8">
     <meta name="viewport" content="width=device-width, initial-scale=1">
     <title>Forum Example</title>
+    <link href="https://fonts.googleapis.com/css?family=Source+Sans+Pro:400,700|Source+Serif+Pro:400" rel="stylesheet">
   </head>
   <body>
     <div id="root"></div>
