+++ examples/forum-relay/src/components/PostPage.js
@@ -1,6 +1,7 @@
 import React from 'react'
 import Relay from 'react-relay'
 import { Link, withRouter } from 'react-router'
+import { StyleSheet, css } from 'aphrodite'
 import { UpdatePostMutation, DeletePostMutation } from '../mutations'
 
 class Post extends React.Component {
@@ -8,19 +9,23 @@ class Post extends React.Component {
     user: React.PropTypes.object,
   }
 
-  handleUpdate(event) {
+  handleUpdate = (event) => {
+    const propName = event.target.dataset.prop
+    const newValue = event.target.innerText
+    const oldValue = this.props.post[propName]
+    if (newValue === oldValue) return
     this.props.relay.commitUpdate(
       new UpdatePostMutation({
         post: this.props.post,
         newPost: {
           // PostGrahpQL expects the prop names of the new values to be prefixed with `new`
-          [`new${capitalizeFirstLetter(event.target.dataset.name)}`]: event.target.innerText,
+          [`new${capitalizeFirstLetter(propName)}`]: newValue,
         },
       })
     )
   }
 
-  handleDelete(event) {
+  handleDelete = (event) => {
     this.props.router.replace('/')
     this.props.relay.commitUpdate(
       new DeletePostMutation({
@@ -34,14 +39,20 @@ class Post extends React.Component {
   // react complains; I chose to ignore it
   render() {
     const { authenticated } = this.context.user
+    const { post } = this.props
     return (
       <div>
         <Link to="/">back to Posts</Link>
-        <h1 data-name="headline" contentEditable={authenticated} onBlur={::this.handleUpdate}>{this.props.post.headline}</h1>
-        <p data-name="body" contentEditable={authenticated} onBlur={::this.handleUpdate}>{this.props.post.body}</p>
-        {authenticated &&
-          <button onClick={::this.handleDelete}>Delete Post</button>
-        }
+        <header>
+          <h1 data-prop="headline" contentEditable={authenticated} onBlur={this.handleUpdate}>{post.headline}</h1>
+          <p>by {post.author.fullName}</p>
+        </header>
+        <p data-prop="body" contentEditable={authenticated} onBlur={this.handleUpdate}>{post.body}</p>
+        <aside>
+          {authenticated &&
+            <button onClick={this.handleDelete}>Delete Post</button>
+          }
+        </aside>
       </div>
     )
   }
@@ -64,6 +75,9 @@ export default Relay.createContainer(withRouter(Post), {
         ${DeletePostMutation.getFragment('post')},
         headline,
         body,
+        author: personByAuthorId {
+          fullName,
+        },
       }
     `,
   },
