+++ examples/forum-relay/src/utils/authDecorator.js
@@ -1,0 +1,96 @@
+import React from 'react'
+import Relay from 'react-relay'
+import jwtDecode from 'jwt-decode'
+import { RelayNetworkLayer, authMiddleware, urlMiddleware } from 'react-relay-network-layer'
+
+const authDecorator = WrappedComponent => {
+  return class AuthDecorator extends React.Component {
+    state = { ...defaultState }
+
+    componentWillMount() {
+      const token = localStorage.getItem('token')
+      if (!token)
+        return
+      else if (isTokenExpired(token))
+        this.clearAuth()
+      else
+        this.setAuth(token)
+    }
+
+    setAuth(token) {
+      setupNetwork(token)
+      this.setState({
+        token,
+        authenticated: true,
+        ...jwtDecode(token),
+      })
+    }
+
+    clearAuth() {
+      localStorage.removeItem('token')
+      setupNetwork(null) // remove token from fetch headers
+      this.setState({
+        ...defaultState
+      })
+    }
+
+    handleLogin = (data) => {
+      fetchToken(data).then(token => {
+        localStorage.setItem('token', token)
+        this.setAuth(token)
+      })
+    }
+
+    handleLogout = () => {
+      this.clearAuth()
+    }
+
+    render() {
+      return (
+        <WrappedComponent
+          user={this.state}
+          handleLogin={this.handleLogin}
+          handleLogout={this.handleLogout}
+          {...this.props}
+        />
+      )
+    }
+  }
+}
+
+const defaultState = {
+  authenticated: false,
+  token: null,
+}
+
+const fetchToken = (data) => {
+  return fetch(AUTH_URL, {
+    method: 'POST',
+    body: JSON.stringify(data),
+    headers: {
+      'Accept': 'application/json',
+      'Content-Type': 'application/json'
+    },
+  })
+  .then(res => res.json())
+  .then(({ token }) => token)
+  .catch(err => console.log(err))
+}
+
+const setupNetwork = (token) => {
+  Relay.injectNetworkLayer(new RelayNetworkLayer([
+    authMiddleware({
+      allowEmptyToken: true,
+      token,
+    }),
+  ], { disableBatchQuery: true }))
+}
+
+const isTokenExpired = (token) => {
+  const tokenPayload = jwtDecode(token)
+  const expiryDate = new Date(tokenPayload.exp * 1000)
+  const currentDate = new Date()
+  return expiryDate < currentDate
+}
+
+export default authDecorator
